Make engine v1 launch, index and trade against a real chain
Gate H — an engine market launched through the app's own code path on a live
chain, traded, indexed, and read back through the production market source —
found two defects that every unit test in the repository had agreed was fine.

The first was a selector. `CONFIG_ABI` declared `threshold` and
`thresholdTokens` as `uint256` where `AgenRuleLib.Config` declares `uint128`,
so `deployMarket` encoded a signature the deployed factory does not answer to.
Byte-parity tests could not see it: both sides were derived from the same wrong
ABI, so they agreed with each other and disagreed only with the chain. Every
launch transaction would have reverted with no reason string. It is now checked
against the compiled artefact's own signature, nested tuples included.

The second was event ordering. The `FeeTaken` handler assumed the `Swap` it
belongs to had already been indexed, which holds for a sell and never for a
buy — a native-quote buy is charged in `beforeSwap`. So every buy was recorded
as having paid nothing, on a market whose whole purpose is the fee it charges.
Fees that arrive early are now parked and claimed by the swap they belong to.

Alongside those: engine launches register in the app instead of being dropped
by a handler that only knew engine-0's event, and a launch is credited to the
wallet that approved the build rather than to whoever sent the transaction;
`readLiveMarket` asks the launching factory's own registry, so an engine market
is no longer looked up in engine-0's and truthfully reported absent; the
approval message has one canonical builder, held byte-for-byte to recorded
vectors through both the server and browser entry points; the market page shows
real trades and the rate the hook emitted, rather than an empty list and the
pool's zero; the deployment script mines the hook salt in-process, so no salt
can go stale between two invocations; and the model providers count their own
failures, because a dead primary vendor is otherwise indistinguishable from a
healthy one.

Also records the production engine treasury, and stops treating broadcast
`run-latest.json` as history: the local rig runs anvil under chain id 4663, so
that file was being overwritten with anvil output under the same path a real
mainnet broadcast writes to, and several committed copies of it were exactly
that. `deployments/robinhood.json` names each live artefact explicitly, which
is what its schema already said to rely on.

+++ packages/market-engine/src/encode.ts
@@ -79,6 +79,24 @@ const ZERO_ADDRESS: Address = "0x0000000000000000000000000000000000000000";
  * whatever builds the calldata is a second definition, and the two would agree until they
  * did not — at which point a creator would have signed a commitment over one configuration
  * and sent a transaction carrying another.
+ *
+ * ## The widths are part of the type, not only of the encoding
+ *
+ * Every integer here has to be the width Solidity declares, and byte-parity is not enough to
+ * establish that. `threshold` and `thresholdTokens` are `uint128` on chain and were `uint256`
+ * here, which is invisible to every hash: the ABI pads both to 32 bytes, so `abi.encode`
+ * produced identical bytes and `RuleLib.vectors.t.sol` passed on every vector.
+ *
+ * What it is not invisible to is the function selector. A selector is `keccak` of the
+ * signature *string*, so a manifest with this tuple nested inside it hashed to
+ * `deployMarket(...(uint256,uint24,uint24)[]...)` while the deployed factory answers to
+ * `...(uint128,uint24,uint24)[]...`. Every launch transaction the app built was addressed to
+ * a function that does not exist, and reverted at the factory with no reason data — after the
+ * review screen, after the signature, after the commitment checks, all of which were correct.
+ *
+ * It was found by `scripts/indexer-proof.sh` putting the app's own calldata on a chain, which
+ * is the only thing that could have found it: it is not reachable from either side alone.
+ * `prepare.test.ts` now asserts this tuple against the compiled contract's own signature.
  */
 export const CONFIG_ABI = [
   {
@@ -94,7 +112,7 @@ export const CONFIG_ABI = [
         type: "tuple[]",
         name: "stages",
         components: [
-          { type: "uint256", name: "threshold" },
+          { type: "uint128", name: "threshold" },
           { type: "uint24", name: "buyFeePpm" },
           { type: "uint24", name: "sellFeePpm" },
         ],
@@ -103,7 +121,7 @@ export const CONFIG_ABI = [
         type: "tuple[]",
         name: "buyTiers",
         components: [
-          { type: "uint256", name: "thresholdTokens" },
+          { type: "uint128", name: "thresholdTokens" },
           { type: "uint24", name: "feePpm" },
         ],
       },
@@ -111,7 +129,7 @@ export const CONFIG_ABI = [
         type: "tuple[]",
         name: "sellTiers",
         components: [
-          { type: "uint256", name: "thresholdTokens" },
+          { type: "uint128", name: "thresholdTokens" },
           { type: "uint24", name: "feePpm" },
         ],
       },
